Reinstall Pods
I needed to do this in order to have my new scheme setup work. It was a
major pain, but seems to be ok. I verified that none of the actual
library files had major changes (I didn’t change the pod versions so
there really shouldn’t have been).